requirements.txt
langchain
langchain-groq
groq
faiss-cpu
sentence-transformers
pypdf
python-dotenv
langchain-community
langchain-huggingface